Откат изменений в ветке 'master'
откат произведенных изменений до коммита bf8728647811fd5810e9a2ad63740dccb58f363f

+++ src/main.c
@@ -1,9 +1,7 @@
 #include "sudoku.h"
 
-// Реализовать считывание поля из файла
-
 int main() {
-    uint8_t** Puzzle;
+    int** Puzzle;
     Puzzle = createPuzzle();
     printPuzzle(Puzzle);
     return 0;
